requirements.txt
pandas==3.0.0
numpy==2.4.2
matplotlib==3.10.8
seaborn==0.13.2
azure-storage-blob==12.26.0
